+++ CHANGES.md
@@ -15,6 +15,9 @@
 * Add regex support to the `pattern` attribute of the `application` and `program` elements of the **Supvisors** rules
   file.
 
+* Update the `RunningFailureHandler` so that `Starter` and `Stopper` actions are all stored before they are actually 
+  triggered.
+
 * The `programs` section has been added in the `application` section of the **Supvisors** rules file.
   All `program` definitions should be placed in this section rather than directly in the `application` section.
   The intention is for the next **Supvisors** version to be able to declare application options in any order.